feat(JOB-187): delete from template removed jobs
